Extracted column_names code from AutoMLTabularTrainingJob for reuse
